% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_fmr}
\alias{get_fmr}
\title{Load Fair Market Rents from HUD API}
\usage{
get_fmr(
  geography,
  state,
  year,
  token = NULL,
  drop_empty_cols = TRUE,
  tibble = TRUE
)
}
\arguments{
\item{geography}{Geography of data to pull. One of "county" or "metro".}

\item{state}{State to pull data for. Specify as an upper or lowercase
two-letter state abbreviation. States supported include the 50 states
plus DC.}

\item{year}{Year to pull data for. Currently, years 2017 to 2020 are
supported.}

\item{token}{HUD API
\href{https://www.huduser.gov/portal/dataset/fmr-api.html}{access token}.
Store your access token in an environment variable named
\code{HUD_API_TOKEN} to allow this function to automatically load it as
needed (see \href{https://github.com/CenterOnBudget/hudapi}{README} for
instructions).}

\item{drop_empty_cols}{If \code{TRUE} (default), drop empty columns in
returned data.}

\item{tibble}{If \code{TRUE} (default), return data as a
\href{https://tibble.tidyverse.org/index.html}{tibble}. Otherwise, return
data as a base data frame.}
}
\value{
A tibble or base data frame with requested data.
}
\description{
\code{get_fmr()} loads Fair Market Rents for counties or metro areas within
a given state from the HUD API.
}
