% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_geo}
\alias{get_geo}
\title{Load geographic data from the HUD API}
\usage{
get_geo(
  geography,
  state = NULL,
  token = get_token(),
  drop_empty_cols = TRUE,
  tibble = TRUE,
  show_url = FALSE
)
}
\arguments{
\item{geography}{Geography of data to pull. One of \code{"state"},
\code{"county"}, or \code{"metro"}.}

\item{state}{If \code{geography} is \code{"county"}, what state to pull data
for. Specify as a two-letter state abbreviation. The 50 states plus DC, AS,
GU, MP, PR, and VI are supported.}

\item{token}{\href{https://www.huduser.gov/portal/dataset/fmr-api.html}{HUD API access token}.
Defaults to environment variable \code{HUD_API_TOKEN}.}

\item{drop_empty_cols}{If \code{TRUE} (default), drop empty columns in
returned data.}

\item{tibble}{If \code{TRUE} (default), return data as a
\href{https://tibble.tidyverse.org/index.html}{tibble}. Otherwise, return
data as a base data frame.}

\item{show_url}{If \code{TRUE}, show URL the request was sent to. Defaults to
\code{FALSE}.}
}
\value{
A tibble or base data frame with requested data.
}
\description{
\code{get_geo()} loads geographic data for states, counties, or metro areas
from the HUD API. This function is particularly useful for looking up county
and METRO codes.
}
