% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_il}
\alias{get_il}
\title{Load Income Limits from HUD API}
\usage{
get_il(
  geography,
  entityid,
  year,
  token = NULL,
  drop_empty_cols = TRUE,
  tibble = TRUE,
  show_url = FALSE
)
}
\arguments{
\item{geography}{Geography of data to pull. One of \code{"state"},
\code{"county"}, or \code{"metro"}.}

\item{entityid}{ID of entity to pull data for. If \code{geography} is
\code{"state"}, ID should be a two-letter state abbreviation. The 50 states
are supported. If \code{geography} is \code{"county"}, ID should be a
10-character county code. If \code{geography} is \code{"metro"}, ID should
be a 16-character METRO code. Using \code{get_geo()} to look up county and
METRO codes is recommended.}

\item{year}{Year to pull data for. Currently, years 2017 to 2020 are
supported.}

\item{token}{HUD API
\href{https://www.huduser.gov/portal/dataset/fmr-api.html}{access token}.
Store your token in env var \code{HUD_API_TOKEN} to pass automatically
(see \href{https://github.com/CenterOnBudget/hudapi}{README} for
instructions).}

\item{drop_empty_cols}{If \code{TRUE} (default), drop empty columns in
returned data.}

\item{tibble}{If \code{TRUE} (default), return data as a
\href{https://tibble.tidyverse.org/index.html}{tibble}. Otherwise, return
data as a base data frame.}

\item{show_url}{If \code{TRUE}, show URL the request was sent to. Defaults to
\code{FALSE}.}
}
\value{
A tibble or base data frame with requested data.
}
\description{
\code{get_il()} loads Income Limits for a given state, county, or
metro area from the HUD API.
}
